pyproject.toml
[project]
name = "pdf2md"
version = "0.3.0"
description = "Convert PDF to Markdown using Gemini Flash 2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "pillow",
    "pdf2image",
    "pymupdf",
    "python-dotenv",
]
